requirements.txt
streamlit
pillow
playsound
